feat; remove badge api and finish rename of GrandPrixApi
