--- src/Backend/Network/GraphApi/BadgeApi.ts
@@ -1,4 +1,0 @@
-// Get all player badges. 
-// Get round scores plus badges.
-// All of a single player's badges for all Grand Prixes
-//
